add real tests for app
